server/requirements.txt
ccxt
pandas
numpy
requests
pymongo
python-dotenv